fix: Tooltip hydration issue
